requirements.txt
opencv-python
numpy
pandas
pyarrow
duckdb
Pillow
scikit-image
scikit-learn
reportlab
matplotlib
jinja2
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.7
openai
anthropic>=0.34.0
tenacity>=8.2
orjson>=3.10
torch
torchvision
scipy
python-dotenv>=1.0
pyyaml
pytest
pytest-cov
ruff
mypy
types-PyYAML
types-requests
python-docx
